Augment `Agent` with available move masking
The availability of the moves is included in the input to the net,
and then the output is constrained according to the availability
mask. This ensures the agents never makes a non-available move and
also allows it to learn the mask (theoretically, the suppression on
the output should not be necessary when the net is perfectly
trained.